+++ plugins/z3_utils/src/subgraph_function_generator.cpp
@@ -1,6 +1,8 @@
-#include "subgraph_function_generator.h"
+#include "plugin_z3_utils.h"
 
 #include "hal_core/netlist/boolean_function.h"
+#include "hal_core/netlist/decorators/boolean_function_net_decorator.h"
+#include "hal_core/netlist/decorators/subgraph_netlist_decorator.h"
 #include "hal_core/netlist/gate.h"
 #include "hal_core/netlist/net.h"
 #include "hal_core/netlist/netlist.h"
@@ -13,305 +15,211 @@ namespace hal
 {
     namespace z3_utils
     {
-        BooleanFunction SubgraphFunctionGenerator::get_function_of_gate(const Gate* gate, const GatePin* out_pin)
+        namespace 
         {
-            if (auto it = m_cache.find({gate->get_id(), out_pin}); it != m_cache.end())
+            Result<BooleanFunction> get_function_of_gate(const Gate* const gate, const GatePin* output_pin, std::map<std::pair<u32, const GatePin*>, BooleanFunction>& cache)
             {
-                return it->second;
-            }
-            BooleanFunction bf = gate->get_boolean_function(out_pin);
-
-            if (bf.is_empty())
-            {
-                log_error("z3_utils", "function of gate {} (type {}) associated with pin {} is empty!", gate->get_name(), gate->get_type()->get_name(), out_pin->get_name());
-                return BooleanFunction();
-            }
-
-            // before replacing input pins with their connected net id, check if the function depends on other output pins
-            auto output_pins = gate->get_type()->get_output_pin_names();
-            while (true)
-            {
-                auto vars = bf.get_variable_names();
-                std::vector<std::string> output_pins_that_are_also_function_inputs;
-                std::set_intersection(vars.begin(), vars.end(), output_pins.begin(), output_pins.end(), std::back_inserter(output_pins_that_are_also_function_inputs));
-
-                if (output_pins_that_are_also_function_inputs.empty())
+                if (auto it = cache.find({gate->get_id(), output_pin}); it != cache.end())
                 {
-                    break;
+                    return OK(it->second);
                 }
 
-                for (auto const& output_pin : output_pins_that_are_also_function_inputs)
+                BooleanFunction bf = gate->get_boolean_function(output_pin);
+
+                std::vector<std::string> input_vars = utils::to_vector(bf.get_variable_names());
+                while (!input_vars.empty())
                 {
-                    const auto gate_func = gate->get_boolean_function(output_pin);
-                    if (auto res = bf.substitute(output_pin, gate_func); res.is_error())
+                    const std::string var = input_vars.back();
+                    input_vars.pop_back();
+
+                    const GatePin* pin = gate->get_type()->get_pin_by_name(var);
+                    if (pin == nullptr)
                     {
-                        log_error("z3_utils",
-                                  "error encountered while substituting variable '{}' with Boolean function of gate '{}' with ID {}:\n{}",
-                                  output_pin,
-                                  gate->get_name(),
-                                  gate->get_id(),
-                                  res.get_error().get());
-                        return BooleanFunction();
+                        return ERR("could not get Boolean function of gate '" + gate->get_name() + "' with ID " + std::to_string(gate->get_id()) + ": failed to get input pin '" + var
+                                    + "' by name");
                     }
-                    else
+
+                    const PinDirection pin_dir = pin->get_direction();
+                    if (pin_dir == PinDirection::input)
                     {
-                        bf = res.get();
+                        const Net* const input_net = gate->get_fan_in_net(var);
+                        if (input_net == nullptr)
+                        {
+                            // if no net is connected, the input pin name cannot be replaced
+                            return ERR("could not get Boolean function of gate '" + gate->get_name() + "' with ID " + std::to_string(gate->get_id()) + ": failed to get fan-in net at pin '"
+                                        + pin->get_name() + "'");
+                        }
+
+                        const auto net_dec = BooleanFunctionNetDecorator(*input_net);
+                        bf                 = bf.substitute(var, net_dec.get_boolean_variable_name());
+                    }
+                    else if ((pin_dir == PinDirection::internal) || (pin_dir == PinDirection::output))
+                    {
+                        BooleanFunction bf_interal = gate->get_boolean_function(var);
+                        if (bf_interal.is_empty())
+                        {
+                            return ERR("could not get Boolean function of gate '" + gate->get_name() + "' with ID " + std::to_string(gate->get_id())
+                                        + ": failed to get Boolean function at output pin '" + pin->get_name() + "'");
+                        }
+
+                        const std::vector<std::string> internal_input_vars = utils::to_vector(bf_interal.get_variable_names());
+                        input_vars.insert(input_vars.end(), internal_input_vars.begin(), internal_input_vars.end());
+
+                        if (auto substituted = bf.substitute(var, bf_interal); substituted.is_error())
+                        {
+                            return ERR_APPEND(substituted.get_error(),
+                                                "could not get Boolean function of gate '" + gate->get_name() + "' with ID " + std::to_string(gate->get_id()) + ": failed to substitute variable '"
+                                                    + var + "' with another Boolean function");
+                        }
+                        else
+                        {
+                            bf = substituted.get();
+                        }
                     }
                 }
+
+                bf = bf.simplify();
+
+                cache.insert({{gate->get_id(), output_pin}, bf});
+                return OK(bf);
             }
 
-            // replace input pins with their connected net id
-            for (const auto& input_pin : gate->get_type()->get_input_pins())
+            Result<z3::expr> get_function_of_net(const std::vector<Gate*>& subgraph_gates, const Net* net, z3::context& ctx, std::map<u32, z3::expr>& net_cache, std::map<std::pair<u32, const GatePin*>, BooleanFunction>& gate_cache)
             {
-                const auto& input_net = gate->get_fan_in_net(input_pin);
-                if (!input_net)
+                if (const auto it = net_cache.find(net->get_id()); it != net_cache.end())
                 {
-                    log_debug("z3_utils", "Pin ({}) has no input net. Gate id: ({})", input_pin->get_name(), gate->get_id());
-                    continue;
+                    return OK(it->second);
                 }
-                bf = bf.substitute(input_pin->get_name(), std::to_string(input_net->get_id()));
+
+                const std::vector<Endpoint*> sources = net->get_sources();
+
+                // net is multi driven
+                if (sources.size() > 1)
+                {
+                    return ERR("cannot get Boolean z3 function of net " + net->get_name() + " with ID " + std::to_string(net->get_id()) + ": net is multi driven.");
+                }
+
+                // net has no source
+                if (sources.empty())
+                {
+                    z3::expr ret = ctx.bv_const(BooleanFunctionNetDecorator(*net).get_boolean_variable_name().c_str(), 1);
+                    net_cache.insert({net->get_id(), ret});
+                    return OK(ret);
+                }
+
+                const Endpoint* src_ep = sources.front();
+
+                if (src_ep->get_gate() == nullptr)
+                {
+                    return ERR("cannot get Boolean z3 function of net " + net->get_name() + " with ID " + std::to_string(net->get_id()) + ": net source is null.");
+                }
+
+                const Gate* src = src_ep->get_gate();
+
+                // source is not in subgraph gates
+                if (std::find(subgraph_gates.begin(), subgraph_gates.end(), src) == subgraph_gates.end())
+                {
+                    z3::expr ret = ctx.bv_const(BooleanFunctionNetDecorator(*net).get_boolean_variable_name().c_str(), 1);
+                    net_cache.insert({net->get_id(), ret});
+                    return OK(ret);
+                }
+
+                const auto& bf_res = get_function_of_gate(src, src_ep->get_pin(), gate_cache);
+                if (bf_res.is_error())
+                {
+                    return ERR_APPEND(bf_res.get_error(), "cannot get Boolean z3 function of net " + net->get_name() + " with ID " + std::to_string(net->get_id()) + ": failed to get function of gate.");
+                }
+                const BooleanFunction bf = bf_res.get();
+
+                std::map<std::string, z3::expr> input_to_expr;
+
+                for (const std::string& in_net_str : bf.get_variable_names())
+                {
+                    const auto in_net_res = BooleanFunctionNetDecorator::get_net_from(src->get_netlist(), in_net_str);
+                    if (in_net_res.is_error())
+                    {
+                        return ERR_APPEND(in_net_res.get_error(), "cannot get Boolean z3 function of net " + net->get_name() + " with ID " + std::to_string(net->get_id()) + ": failed to reconstruct input net from variable " + in_net_str + ".");
+                    }
+                    const auto in_net = in_net_res.get();
+
+                    const auto in_bf_res = get_function_of_net(subgraph_gates, in_net, ctx, net_cache, gate_cache);
+                    if (in_bf_res.is_error())
+                    {
+                        // NOTE since this can lead to a deep recursion we do not append the error and instead only propagate it.
+                        return in_bf_res;
+                    }
+                    const auto in_bf = in_bf_res.get();
+
+                    input_to_expr.insert({in_net_str, in_bf});
+                }
+
+                z3::expr ret = z3_utils::to_z3(bf, ctx, input_to_expr).simplify();
+                net_cache.insert({net->get_id(), ret});
+
+                return OK(ret);
             }
+        
+            Result<z3::expr> get_subgraph_z3_function_internal(const std::vector<Gate*>& subgraph_gates, const Net* net, z3::context& ctx, std::map<u32, z3::expr>& net_cache, std::map<std::pair<u32, const GatePin*>, BooleanFunction>& gate_cache) 
+            {
+                // check validity of subgraph_gates
+                if (subgraph_gates.empty())
+                {
+                    return ERR("could not get subgraph z3 function of net '" + net->get_name() + "' with ID " + std::to_string(net->get_id()) + ": subgraph contains no gates");
+                }
+                else if (std::any_of(subgraph_gates.begin(), subgraph_gates.end(), [](const Gate* g) { return g == nullptr; }))
+                {
+                    return ERR("could not get subgraph z3 function of net '" + net->get_name() + "' with ID " + std::to_string(net->get_id()) + ": subgraph contains a gate that is a 'nullptr'");
+                }
+                else if (net == nullptr)
+                {
+                    return ERR("could not get subgraph z3 function: net is a 'nullptr'");
+                }
+                else if (net->get_num_of_sources() > 1)
+                {
+                    return ERR("could not get subgraph z3 function of net '" + net->get_name() + "' with ID " + std::to_string(net->get_id()) + ": net has more than one source");
+                }
+                else if (net->is_global_input_net())
+                {
+                    const auto net_dec = BooleanFunctionNetDecorator(*net);
+                    return OK(ctx.bv_const(net_dec.get_boolean_variable_name().c_str(), 1));
+                }
+                else if (net->get_num_of_sources() == 0)
+                {
+                    return ERR("could not get subgraph function of net '" + net->get_name() + "' with ID " + std::to_string(net->get_id()) + ": net has no sources");
+                }
 
-            m_cache.emplace(std::make_tuple(gate->get_id(), out_pin), bf);
+                return get_function_of_net(subgraph_gates, net, ctx, net_cache, gate_cache);
+            }
+        
+        }  // namespace
+        
+        Result<z3::expr> get_subgraph_z3_function(const std::vector<Gate*>& subgraph_gates, const Net* subgraph_output, z3::context& ctx)
+        {
+            std::map<u32, z3::expr> net_cache;
+            std::map<std::pair<u32, const GatePin*>, BooleanFunction> gate_cache;
 
-            return bf;
+            return get_subgraph_z3_function_internal(subgraph_gates, subgraph_output, ctx, net_cache, gate_cache);
         }
 
-        void SubgraphFunctionGenerator::get_subgraph_z3_function(const Net* output_net,
-                                                                 const std::vector<Gate*> subgraph_gates,
-                                                                 z3::context& ctx,
-                                                                 z3::expr& result,
-                                                                 std::unordered_set<u32>& input_net_ids)
+        Result<std::vector<z3::expr>> get_subgraph_z3_functions(const std::vector<Gate*>& subgraph_gates, const std::vector<Net*> subgraph_outputs, z3::context& ctx)
         {
-            /* check validity of subgraph_gates */
-            if (subgraph_gates.empty())
-            {
-                log_error("z3_utils", "parameter 'subgraph_gates' is empty");
-            }
-            if (std::any_of(subgraph_gates.begin(), subgraph_gates.end(), [](auto& g) { return g == nullptr; }))
-            {
-                log_error("z3_utils", "parameter 'subgraph_gates' contains a nullptr");
-            }
-            if (output_net->get_num_of_sources() != 1)
-            {
-                log_error("z3_utils", "target net has 0 or more than 1 sources.");
-            }
+            std::map<u32, z3::expr> net_cache;
+            std::map<std::pair<u32, const GatePin*>, BooleanFunction> gate_cache;
 
-            std::queue<Net*> q;
+            std::vector<z3::expr> results;
 
-            input_net_ids.clear();
-
-            // prepare queue and result with initial gate
+            for (const auto& net : subgraph_outputs)
             {
-                auto source     = output_net->get_sources()[0];
-                auto start_gate = source->get_gate();
-
-                // Check wether start gate is in the subgraph gates
-                if (std::find(subgraph_gates.begin(), subgraph_gates.end(), start_gate) == subgraph_gates.end())
+                const auto res = get_subgraph_z3_function_internal(subgraph_gates, net, ctx, net_cache, gate_cache);
+                if (res.is_error())
                 {
-                    result = ctx.bv_const(std::to_string(output_net->get_id()).c_str(), 1);
-                    return;
+                    return ERR_APPEND(res.get_error(), "unable to generate subgraph functions: failed to generate function for net " + net->get_name() + " with ID " + std::to_string(net->get_id()) + ".");
                 }
 
-                auto f = get_function_of_gate(start_gate, source->get_pin());
-
-                for (auto id_string : f.get_variable_names())
-                {
-                    input_net_ids.insert(std::stoi(id_string));
-                }
-
-                result = f.to_z3(ctx);
-
-                for (auto& n : start_gate->get_fan_in_nets())
-                {
-                    q.push(n);
-                }
+                results.push_back(res.get());
             }
 
-            while (!q.empty())
-            {
-                auto n = q.front();
-                q.pop();
-
-                if (n->get_num_of_sources() != 1)
-                {
-                    //log_error("z3_utils", "net has 0 or more than 1 sources. not expanding the function here");
-                    continue;
-                }
-
-                if (input_net_ids.find(n->get_id()) == input_net_ids.end())
-                {
-                    // log_info("z3_utils", "net {} (id {}) is not an input to the function", n->get_name(), n->get_id());
-                    continue;
-                }
-
-                auto src      = n->get_sources()[0];
-                auto src_gate = src->get_gate();
-
-                if (std::find(subgraph_gates.begin(), subgraph_gates.end(), src_gate) != subgraph_gates.end())
-                {
-                    auto f = get_function_of_gate(src_gate, src->get_pin());
-
-                    // substitute function into z3 expression
-                    z3::expr_vector from_vec(ctx);
-                    z3::expr_vector to_vec(ctx);
-
-                    z3::expr from = ctx.bv_const(std::to_string(n->get_id()).c_str(), 1);
-                    z3::expr to   = f.to_z3(ctx);
-
-                    from_vec.push_back(from);
-                    to_vec.push_back(to);
-
-                    result = result.substitute(from_vec, to_vec);
-
-                    // replace the net id which was substituted with all input ids of the substituted function
-                    input_net_ids.erase(n->get_id());
-                    for (auto id_string : f.get_variable_names())
-                    {
-                        input_net_ids.insert(std::stoi(id_string));
-                    }
-
-                    // add input nets of gate to queue
-                    for (auto& sn : src_gate->get_fan_in_nets())
-                    {
-                        q.push(sn);
-                    }
-                }
-            }
-        }
-
-        RecursiveSubgraphFunctionGenerator::RecursiveSubgraphFunctionGenerator(z3::context& ctx, const std::vector<Gate*>& subgraph_gates) : m_ctx(&ctx), m_subgraph_gates(subgraph_gates)
-        {
-        }
-
-        BooleanFunction RecursiveSubgraphFunctionGenerator::get_function_of_gate(const Gate* gate, const GatePin* out_pin)
-        {
-            if (auto it = m_cache.find({gate->get_id(), out_pin}); it != m_cache.end())
-            {
-                return it->second;
-            }
-            BooleanFunction bf = gate->get_boolean_function(out_pin);
-
-            if (bf.is_empty())
-            {
-                log_error("z3_utils", "function of gate {} (type {}) associated with pin {} is empty!", gate->get_name(), gate->get_type()->get_name(), out_pin->get_name());
-                return BooleanFunction();
-            }
-
-            // TODO should this also take internal pins into account?
-            // before replacing input pins with their connected net id, check if the function depends on other output pins
-            std::vector<std::string> output_pins = gate->get_type()->get_output_pin_names();
-            while (true)
-            {
-                auto vars = bf.get_variable_names();
-                std::vector<std::string> output_pins_that_are_also_function_inputs;
-                std::set_intersection(vars.begin(), vars.end(), output_pins.begin(), output_pins.end(), std::back_inserter(output_pins_that_are_also_function_inputs));
-
-                if (output_pins_that_are_also_function_inputs.empty())
-                {
-                    break;
-                }
-
-                for (auto const& output_pin : output_pins_that_are_also_function_inputs)
-                {
-                    const auto gate_func = gate->get_boolean_function(output_pin);
-                    if (auto res = bf.substitute(output_pin, gate_func); res.is_error())
-                    {
-                        log_error("z3_utils",
-                                  "error encountered while substituting variable '{}' with Boolean function of gate '{}' with ID {}:\n{}",
-                                  output_pin,
-                                  gate->get_name(),
-                                  gate->get_id(),
-                                  res.get_error().get());
-                        return BooleanFunction();
-                    }
-                    else
-                    {
-                        bf = res.get();
-                    }
-                }
-            }
-
-            m_cache.emplace(std::make_tuple(gate->get_id(), out_pin), bf);
-
-            return bf;
-        }
-
-        z3::expr RecursiveSubgraphFunctionGenerator::get_function_of_net(const Net* net, z3::context& ctx, const std::vector<Gate*>& subgraph_gates)
-        {
-            if (m_expr_cache.find(net) != m_expr_cache.end())
-            {
-                return m_expr_cache.at(net);
-            }
-
-            const std::vector<Endpoint*> sources = net->get_sources();
-
-            // net is multi driven
-            if (sources.size() > 1)
-            {
-                log_error("z3_utils", "Cannot handle multi driven nets! Encountered at net {}.", net->get_id());
-                return ctx.bv_const("ERROR", 1);
-            }
-
-            // net has no source
-            if (sources.empty())
-            {
-                z3::expr ret = ctx.bv_const(std::to_string(net->get_id()).c_str(), 1);
-                m_expr_cache.insert({net, ret});
-                return ret;
-            }
-
-            const Endpoint* src_ep = sources.front();
-
-            if (src_ep->get_gate() == nullptr)
-            {
-                log_error("z3_utils", "Gate at source for net {} is null.", net->get_id());
-            }
-
-            const Gate* src = src_ep->get_gate();
-
-            // source is not in subgraph gates
-            if (std::find(subgraph_gates.begin(), subgraph_gates.end(), src) == subgraph_gates.end())
-            {
-                z3::expr ret = ctx.bv_const(std::to_string(net->get_id()).c_str(), 1);
-                m_expr_cache.insert({net, ret});
-                return ret;
-            }
-
-            const BooleanFunction bf = get_function_of_gate(src, src_ep->get_pin());
-
-            std::map<std::string, z3::expr> pin_to_expr;
-
-            for (const std::string& pin_name : bf.get_variable_names())
-            {
-                Net* in_net = src->get_fan_in_net(pin_name);
-                if (in_net == nullptr)
-                {
-                    log_error("z3_utils", "Cannot find in_net at pin {} of gate {}!", pin_name, src->get_id());
-                }
-
-                pin_to_expr.insert({pin_name, get_function_of_net(in_net, ctx, subgraph_gates)});
-            }
-
-            z3::expr ret = bf.to_z3(ctx, pin_to_expr);
-            m_expr_cache.insert({net, ret});
-            return ret;
-        }
-
-        void RecursiveSubgraphFunctionGenerator::get_subgraph_z3_function_recursive(const Net* net, z3::expr& result)
-        {
-            /* check validity of subgraph_gates */
-            if (m_subgraph_gates.empty())
-            {
-                log_error("z3_utils", "parameter 'subgraph_gates' is empty");
-            }
-            if (std::any_of(m_subgraph_gates.begin(), m_subgraph_gates.end(), [](auto& g) { return g == nullptr; }))
-            {
-                log_error("z3_utils", "parameter 'subgraph_gates' contains a nullptr");
-            }
-
-            result = get_function_of_net(net, *m_ctx, m_subgraph_gates);
-            return;
+            return OK(results);
         }
 
     }    // namespace z3_utils