+++ src/python_bindings/python_bindings.cpp
@@ -67,6 +67,10 @@ namespace hal
 
         smt_init(m);
 
+        boolean_function_net_decorator_init(m);
+
+        subgraph_netlist_decorator_init(m);
+
 #ifndef PYBIND11_MODULE
         return m.ptr();
 #endif    // PYBIND11_MODULE