+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   * added (limited) support for 'defparam' statements to Verilog parser
   * added support for pin assignments by order instead of name to Verilog and VHDL parser
 * miscellaneous
+  * added functions `get_pin_names`, `get_input_pins`, `get_input_pin_names`, `get_output_pins`, and `get_output_pin_names` to class `Module`
   * added function `BooleanFunction::get_truth_table_as_string` that returns the truth table of a Boolean function as a formatted string
   * added missing GND, VCC, and RAM gate types to the `ICE40ULTRA` gate library
   * added Python bindings for the HAL project manager
@@ -21,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   * fixed segfault that sometimes occurred when deleting a module
   * fixed saving absolute paths for Python files and not copying them to the new project folder when using `Save as...`
   * fixed some project manager bugs related to inaccessible files
+  * fixed missing Python binding for `GateType::get_pin_groups`
 
 ## [4.0.1] - 2022-10-24 15:33:15+02:00 (urgency: medium)
 * **WARNING:** this release breaks multiple APIs, please make sure to adjust your code accordingly.